+++ modules/help_browser/src/c/dllMain.c
@@ -24,6 +24,7 @@
 // LICENCE_BLOCK_END
 //=============================================================================
 #include <Windows.h>
+#include <QtCore/QtGlobal>
 //=============================================================================
 #include "nlsConfig.h"
 //=============================================================================
@@ -31,6 +32,14 @@
 #pragma comment(lib, CAT_3_STRINGS("boost_filesystem-", BOOST_TARGET, ".lib"))
 #pragma comment(lib, CAT_3_STRINGS("boost_thread-", BOOST_TARGET, ".lib"))
 #pragma comment(lib, "shlwapi.lib") // AllocConsole
+//=============================================================================
+#if QT_VERSION >= QT_VERSION_CHECK(6, 0, 0)
+#pragma comment(lib, "Qt6Sql.lib")
+#pragma comment(lib, "Qt6Core.lib")
+#else
+#pragma comment(lib, "Qt5Sql.lib")
+#pragma comment(lib, "Qt5Core.lib")
+#endif
 //=============================================================================
 int WINAPI
 DllMain(HINSTANCE hInstance, DWORD reason, PVOID pvReserved)